+++ src/application/useCases/countRecipientNotificationsMyWay.ts
@@ -1,0 +1,26 @@
+import { Injectable } from '@nestjs/common';
+import { NotificationsRepository } from '@application/repositories/notificationsRepository';
+
+interface CountRecipientNotificationRequest {
+  recipientId: string;
+}
+
+interface CountRecipientNotificationResponse {
+  count: number;
+}
+
+@Injectable()
+export class CountRecipientNotifications {
+  constructor(private notificationsRepository: NotificationsRepository) {}
+
+  async execute(
+    request: CountRecipientNotificationRequest,
+  ): Promise<CountRecipientNotificationResponse> {
+    const { recipientId } = request;
+    const count = await this.notificationsRepository.countManyByRecipientId(
+      recipientId,
+    );
+
+    return { count };
+  }
+}